remote_server/requirements.txt
# HTTP 服务器（InternNav AgentServer）
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.28.0

# 图像处理
numpy>=1.20.0
Pillow>=8.0.0

# PyTorch - 云端推理必需
torch>=2.0.0
torchvision>=0.15.0

# transformers（模型加载）
transformers>=4.30.0

# 旧版 gRPC 依赖（可选，兼容旧客户端）
# grpcio>=1.50.0
# grpcio-tools>=1.50.0
# protobuf>=4.21.0
